app/requirements.txt
django-environ<=0.11.2
djangorestframework<=3.15.2
djangorestframework-simplejwt[crypto]<=5.3.1
drf-spectacular[sidecar]==0.27.2
pytest-django<=4.9.0
